Add initial Github Actions CI workflows.
The plan is to support the three most recent stable DMD compilers and the equivalent LDC and GDC compilers. This PR only tests two DMD and one LDC stable compilers, as well as the current pre-release and master branches. We may need to fix the commented-out compiler versions, or we may not -- by the time we have a stable camino release they may be unsupported.